Refactor: Move stripe and span structure to CacheDefs
